update ResponseEntities on controller

+++ src/main/java/pl/postek/webservice/notes/controller/NoteController.java
@@ -1,9 +1,13 @@
 package pl.postek.webservice.notes.controller;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.postek.webservice.notes.exception.NoteNotFoundException;
 import pl.postek.webservice.notes.model.Note;
 import pl.postek.webservice.notes.service.NoteService;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -16,27 +20,45 @@ public class NoteController {
     }
 
     @GetMapping("/allNotes")
-    public List<Note> getAllNotes() {
-       return service.getAllNotes();
+    public ResponseEntity<List<Note>> getAllNotes() {
+        List<Note> allNotes = service.getAllNotes();
+        return new ResponseEntity<>(allNotes, HttpStatus.OK);
     }
 
     @PostMapping("/addNote")
-    public Note saveNote(@RequestBody Note note) {
-        return service.saveNote(note);
+    public ResponseEntity<Object> saveNote(@RequestBody @Valid Note note) {
+        Note savedNote = service.saveNote(note);
+        return new ResponseEntity<>("Note is created successfully with Id = " + savedNote.getId(),
+                HttpStatus.CREATED);
     }
 
     @PutMapping("/updateNote/{id}")
-    public Note updateNote(@PathVariable Long id, @RequestBody Note note) {
-        return service.updateNote(id, note);
+    public ResponseEntity<Object> updateNote(@PathVariable Long id, @RequestBody @Valid Note note) {
+        if (service.noteExistsById(id)) {
+            Note savedNote = service.updateNote(id, note);
+            return new ResponseEntity<>(String.format("Note with id = %d is updated successfully", savedNote.getId()),
+                    HttpStatus.OK);
+        } else
+            throw new NoteNotFoundException(String.format("Note with id = %d not found", id));
     }
 
     @DeleteMapping("/deleteNote/{id}")
-    public void deleteNote(@PathVariable Long id) {
-        service.deleteNote(id);
+    public ResponseEntity<Object> deleteNote(@PathVariable Long id) {
+        if (service.noteExistsById(id)) {
+            service.deleteNote(id);
+            return new ResponseEntity<>(String.format("Note with id = %d is deleted successfully", id),
+                    HttpStatus.OK);
+        } else {
+            throw new NoteNotFoundException(String.format("Note with id = %d not found", id));
+        }
     }
 
     @GetMapping("/revision/{id}")
-    public void lastChangeRevision(@PathVariable Long id) {
-        service.getInfoLastChangeRevision(id);
+    public ResponseEntity<Object> lastChangeRevision(@PathVariable Long id) {
+        if (service.noteExistsById(id)) {
+            service.getInfoLastChangeRevision(id);
+            return new ResponseEntity<>("service.getInfoLastChangeRevision(id)", HttpStatus.OK);
+        }
+        throw new NoteNotFoundException(String.format("Note with id = %d not found", id));
     }
 }
